Compare player guess with hidden code on submit.
Current solution:
Iterate through the guess array and compare with a copy of the
hidden code object to check how many exact matches there are and
mark them so they are not counted again for our next loop.
Then iterate through the guess array again to check how many close
matches there are (location not exact) - the remaining numbers
in the array that are present in the object but not at the right index.

Requires further testing & probably a refactoring.